Create helmets.script
In the original Stalker Call of Pripyat game there is only one inventory slot for a helmet. This means that it's possible to either equip a combat helmet or a gasmask, for example, but not both at the same time.
In Anomaly mod there's an built-in config file that makes it possible to create new items by combining two other items (basically, doing drag-and-drop in the in-game inventory menu). This script is a part of the system that makes it possible to create a new 'combined' helmet (a combat helmet plus any of the present gasmasks, done using vanilla config files) and then split the 'combined' helmet back into a helmet and a mask based on what items were used.